// src/types.ts
export type BackgroundColors =
  | "Blue"
  | "Gold"
  | "Lime"
  | "Pink"
  | "Salmon"
  | "Turquoise"
  | "White";
export type BaseColor =
  | "Pink"
  | "Peach"
  | "Brown"
  | "White"
  | "Gold"
  | "Diamond"
  | "Lime"
  | "Black";
export type HoodieColor = "Red" | "Orange" | "Purple";
export type ISpecialType = "None" | "Skull Face" | "TV Head" | "Giant Eye";
export type TailTypes = "Short" | "Normal" | "Curled" | "Wiggles";
export type IFrillType =
  | "Puffy"
  | "Teenie Tiny"
  | "Short"
  | "Wide"
  | "Thin"
  | "Wiggles";
export type IFaceType =
  | "Anger"
  | "Clout Goggles"
  | "Derp"
  | "Lash"
  | "Oval"
  | "Troll"
  | "Visor"
  | "Bored"
  | "Disapproval";
export type IMouthType =
  | "Bleh"
  | "Blunt"
  | "Happy mouth"
  | "owo"
  | "Plain"
  | "Smile"
  | "Smirk"
  | "Smol frown"
  | "Fangs";
export type IHeadType =
  | "Bald"
  | "Tuft"
  | "Side"
  | "Cowboy Hat"
  | "Crown"
  | "Halo"
  | "Octopus"
  | "Party Hat"
  | "Spin Beanie"
  | "Beanie"
  | "Birdie";
export type IArmType = "Claws" | "Paws";
export type IAccessoriesType =
  | "None"
  | "Bow Tie"
  | "Heart"
  | "Neck Tie"
  | "Rainbow Cape"
  | "Flamingo"
  | "Hoodie"
  | "Floaties";